Added an additional boolean to the Asset constructor, isDefensive. Using isOffensive and isDefensive will allow us to quickly identify weapons, armor/wearable equipment, and healing items. isOffensive and isDefensive default to false, so you don't have to set them both for every asset. You only need to enter one (true) for a weapon, and (false, true) for armor
Added a vector of assets called _Equipment to
store the player's equipped weapon and armor.

The UseAsset function will check the boolean flags
of the asset and either equip the weapon/armor to
the corresponding slot in _Equipment, or use the
item to heal. Weapons/Armor when switched to will
move the current weapon/armor to the _assets
vector.